requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
langchain==0.1.0
langchain-text-splitters==0.0.1
pinecone-client==3.0.0
qdrant-client==1.7.0
sentence-transformers==2.2.2
GitPython==3.1.40
PyGithub==2.1.1
python-gitlab==4.2.0
tiktoken==0.5.2
